integrate queue progress into media lists + series details

+++ composeApp/src/androidMain/kotlin/com/dnfapps/arrmatey/navigation/NavigationDestinations.kt
@@ -12,11 +12,11 @@ sealed interface HomeTab : NavKey {
 
 sealed interface ArrScreen : NavKey {
     data object Library: ArrScreen
-    data class Details(val id: Int): ArrScreen
+    data class Details(val id: Long): ArrScreen
     data class Preview<T>(val item: T): ArrScreen
     data class Search(val query: String = ""): ArrScreen
-    data class MovieReleases(val movieId: Int): ArrScreen
-    data class SeriesRelease(val seriesId: Int? = null, val seasonNumber: Int? = null, val episodeId: Long? = null): ArrScreen
+    data class MovieReleases(val movieId: Long): ArrScreen
+    data class SeriesRelease(val seriesId: Long? = null, val seasonNumber: Int? = null, val episodeId: Long? = null): ArrScreen
 }
 
 sealed interface SettingsScreen : NavKey {
