+++ shared/src/commonMain/kotlin/com/dnfapps/arrmatey/api/arr/model/ArrMovie.kt
@@ -28,7 +28,7 @@ import kotlin.time.Instant
 
 @Serializable
 data class ArrMovie(
-    override val id: Int? = null,
+    override val id: Long? = null,
     override val title: String,
     override val originalLanguage: Language,
     override val year: Int,
