+++ shared/src/commonMain/kotlin/com/dnfapps/arrmatey/api/arr/model/CommandPayload.kt
@@ -5,11 +5,11 @@ import kotlinx.serialization.Serializable
 @Serializable
 sealed class CommandPayload(val name: String) {
     @Serializable
-    data class Movie(val movieIds: List<Int>): CommandPayload("MoviesSearch")
+    data class Movie(val movieIds: List<Long>): CommandPayload("MoviesSearch")
     @Serializable
-    data class Series(val seriesId: Int): CommandPayload("SeriesSearch")
+    data class Series(val seriesId: Long): CommandPayload("SeriesSearch")
     @Serializable
-    data class Season(val seriesId: Int, val seasonNumber: Int): CommandPayload("SeasonSearch")
+    data class Season(val seriesId: Long, val seasonNumber: Int): CommandPayload("SeasonSearch")
     @Serializable
     data class Episode(val episodeIds: List<Long>): CommandPayload("EpisodeSearch")
     @Serializable
